Separate errbuf from curl::Easy_t
Adding error_buffer as a part of curl::Easy_t go against the golden rule
of C++:

 > You only pay for what you use

It also unnecessarily, makes curl::Easy_t and curl::Easy_ref_t more
complex and harder to use, especially in Multi_t.

It also brings about the problem of how to store error_buffer during
call to Multi_t.

Before this commit, CURLOPT_PRIVATE is used for this, however this
prevent client code from ever using this option to store the data they'd
interested in.

So, at the end of the day, this is removed.